pyproject.toml
[project]
name = "universal-text-extractor"
version = "1.0.0"
description = "A comprehensive tool for extracting text from various document formats with OCR capabilities"
authors = [
    {name = "Universal Text Extractor Team"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["text-extraction", "ocr", "document-processing", "pdf", "streamlit"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Text Processing",
]

dependencies = [
    # Core web framework
    "streamlit>=1.45.1",
    
    # Image processing and OCR
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.3",
    
    # PDF processing
    "PyPDF2>=3.0.1",
    "PyMuPDF>=1.23.0",
    
    # Office document processing
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    "docx2txt>=0.8",
    
    # Web and markup processing
    "beautifulsoup4>=4.12.2",
    "defusedxml>=0.7.1",
    
    # Rich text and other formats
    "striprtf>=0.0.22",
    "odfpy>=1.4.1",
    "ebooklib>=0.18",
    "extract-msg>=0.41.1",
    
    # Data processing and export
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    
    # Image processing enhancements
    "opencv-python-headless>=4.8.0",
    "scikit-image>=0.21.0",
    
    # Additional utilities
    "python-magic>=0.4.27",
    "chardet>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

docker = [
    "gunicorn>=21.2.0",
]

[project.urls]
Homepage = "https://github.com/universal-text-extractor/universal-text-extractor"
Documentation = "https://github.com/universal-text-extractor/universal-text-extractor/blob/main/README.md"
Repository = "https://github.com/universal-text-extractor/universal-text-extractor.git"
Issues = "https://github.com/universal-text-extractor/universal-text-extractor/issues"

[project.scripts]
text-extractor = "app:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true